refactor: remove custom llm client
